+++ z_qrcn_test.go
@@ -16,6 +16,7 @@
 package goracle_test
 
 import (
+	"context"
 	"strings"
 	"testing"
 
@@ -24,7 +25,9 @@ import (
 )
 
 func TestQRCN(t *testing.T) {
-	conn, err := goracle.DriverConn(testDb)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	conn, err := goracle.DriverConn(ctx, testDb)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -43,7 +46,7 @@ func TestQRCN(t *testing.T) {
 	s, err := conn.NewSubscription("subscr", cb)
 	if err != nil {
 		errS := errors.Cause(err).Error()
-		if strings.Contains(errS, "ORA-29970:") {
+		if strings.Contains(errS, "ORA-29970:") || strings.Contains(errS, "ORA-65131:") {
 			t.Skip(err.Error())
 		} else if strings.Contains(errS, "ORA-29972:") {
 			t.Log("See \"https://docs.oracle.com/database/121/ADFNS/adfns_cqn.htm#ADFNS553\"")
@@ -67,7 +70,7 @@ func TestQRCN(t *testing.T) {
 	t.Log("--- Registrations ---")
 	for rows.Next() {
 		var regID, table string
-		if err := rows.Scan(&regID,&table); err != nil {
+		if err := rows.Scan(&regID, &table); err != nil {
 			t.Error(err)
 			break
 		}